Renamed 'serial_cli' and 'fc_main'.
