+++ Makefile
@@ -525,13 +525,13 @@ COMMON_SRC = \
             fc/fc_init.c \
             fc/fc_dispatch.c \
             fc/fc_hardfaults.c \
-            fc/fc_main.c \
+            fc/fc_core.c \
             fc/fc_msp.c \
             fc/fc_tasks.c \
             fc/rc_controls.c \
             fc/rc_curves.c \
             fc/runtime_config.c \
-            fc/serial_cli.c \
+            fc/cli.c \
             flight/altitudehold.c \
             flight/failsafe.c \
             flight/imu.c \
@@ -702,7 +702,7 @@ SPEED_OPTIMISED_SRC := $(SPEED_OPTIMISED_SRC) \
 
 SIZE_OPTIMISED_SRC := $(SIZE_OPTIMISED_SRC) \
             drivers/serial_escserial.c \
-            io/serial_cli.c \
+            io/cli.c \
             io/serial_4way.c \
             io/serial_4way_avrootloader.c \
             io/serial_4way_stk500v2.c \
