+++ src/main/fc/cli.c
@@ -69,7 +69,7 @@ uint8_t cliMode = 0;
 #include "fc/config.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
-#include "fc/serial_cli.h"
+#include "fc/cli.h"
 
 #include "flight/failsafe.h"
 #include "flight/imu.h"
