+++ src/main/fc/fc_core.c
@@ -43,7 +43,7 @@
 #include "fc/rc_controls.h"
 #include "fc/rc_curves.h"
 #include "fc/runtime_config.h"
-#include "fc/serial_cli.h"
+#include "fc/cli.h"
 
 #include "msp/msp_serial.h"
