+++ src/main/fc/fc_msp.c
@@ -49,7 +49,7 @@
 #include "drivers/serial_escserial.h"
 
 #include "fc/config.h"
-#include "fc/fc_main.h"
+#include "fc/fc_core.h"
 #include "fc/fc_msp.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
