+++ src/main/fc/fc_tasks.c
@@ -36,10 +36,10 @@
 #include "fc/config.h"
 #include "fc/fc_msp.h"
 #include "fc/fc_tasks.h"
-#include "fc/fc_main.h"
+#include "fc/fc_core.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
-#include "fc/serial_cli.h"
+#include "fc/cli.h"
 #include "fc/fc_dispatch.h"
 
 #include "flight/pid.h"
