+++ src/main/flight/pid.c
@@ -28,7 +28,7 @@
 #include "common/maths.h"
 #include "common/filter.h"
 
-#include "fc/fc_main.h"
+#include "fc/fc_core.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
